+++ java/src/main/java/io/cucumber/messages/ndjson/Serializer.java
@@ -1,20 +1,14 @@
 package io.cucumber.messages.ndjson;
 
-import io.cucumber.messages.MessageToNdjsonWriter;
-import io.cucumber.messages.types.Envelope;
-
-import java.io.IOException;
 import java.io.Writer;
-
-import static java.util.Objects.requireNonNull;
 
 /**
  * Serializes a message to a single line of JSON.
  */
-public final class Serializer implements MessageToNdjsonWriter.Serializer {
+public interface Serializer<T> {
 
     /**
-     * Serialize a message to single line of JSON and write it to the
+     * Serialize an object to single line of JSON and write it to the
      * given {@code writer}.
      *
      * <ul>
@@ -27,12 +21,6 @@ public final class Serializer implements MessageToNdjsonWriter.Serializer {
      *
      * @param writer to write to
      * @param value  to serialize
-     * @throws IOException if anything goes wrong
      */
-    @Override
-    public void writeValue(Writer writer, Envelope value) throws IOException {
-        requireNonNull(writer);
-        requireNonNull(value);
-        Jackson.OBJECT_MAPPER.writeValue(writer, value);
-    }
+    void writeValue(Writer writer, T value);
 }