+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Make Jackson an optional non-shaded dependency
+- Support Jackson 2 and Jackson 3
+- Add `module-info.java`
+- Use JSpecify for nullability
 
 ## [0.4.1] - 2026-03-02
 ### Changed